split reader menu into multiple components
